backend/requirements.txt
Flask==2.2.5
flask-socketio==5.3.2
python-dotenv==1.0.0
redis==4.5.5
eventlet==0.34.1
pytest==7.4.0
requests==2.31.0
